fixes #2546. splits CrudTrait into multiple traits
